Move SnapshotNode and SnapshotConfig
