+++ core/dbt/artifacts/resources/__init__.py
@@ -23,6 +23,7 @@ from dbt.artifacts.resources.v1.manifest_nodes import (
     SingularTestNode,
     TestMetadata,
     GenericTestNode,
+    SnapshotNode,
 )
 from dbt.artifacts.resources.v1.documentation import Documentation
 from dbt.artifacts.resources.v1.exposure import (
@@ -78,4 +79,5 @@ from dbt.artifacts.resources.v1.config import (
     ModelConfig,
     SeedConfig,
     TestConfig,
+    SnapshotConfig,
 )
